Extract conversion to ActivityDeclaration into its own file
