+++ lib/Activities/docu/src/sources.h
@@ -1,3 +1,5 @@
+#pragma once
+
 #include "clang/Tooling/CompilationDatabase.h"
 
 #include <memory>
